refactor: Implement PR feedback.
