+++ azure-boards-common/src/main/java/com/synopsys/integration/azure/boards/common/oauth/AzureAuthorizationCodeFlow.java
@@ -1,5 +1,5 @@
 /**
- * blackduck-alert
+ * azure-boards-common
  *
  * Copyright (c) 2020 Synopsys, Inc.
  *
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.azure.boards.oauth;
+package com.synopsys.integration.azure.boards.common.oauth;
 
 import com.google.api.client.auth.oauth2.AuthorizationCodeFlow;
 import com.google.api.client.auth.oauth2.AuthorizationCodeTokenRequest;
