+++ azure-boards-common/src/main/java/com/synopsys/integration/azure/boards/common/oauth/AzureTokenResponse.java
@@ -1,5 +1,5 @@
 /**
- * blackduck-alert
+ * azure-boards-common
  *
  * Copyright (c) 2020 Synopsys, Inc.
  *
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.channel.azure.boards.oauth;
+package com.synopsys.integration.azure.boards.common.oauth;
 
 import com.google.api.client.auth.oauth2.TokenResponse;
 import com.google.api.client.json.JsonString;
@@ -29,7 +29,7 @@ import com.google.api.client.util.Key;
 public class AzureTokenResponse extends TokenResponse {
 
     // Azure OAuth returns the expires_in as a string not a numeric value.
-    // need to override and create a customer TokenResponse to use the JSonString annotation.
+    // need to override and create a custom TokenResponse to use the JsonString annotation.
     @Key("expires_in")
     @JsonString
     private Long expiresInSeconds;
