+++ src/main/java/com/synopsys/integration/alert/channel/azure/boards/descriptor/AzureBoardsGlobalUIConfig.java
@@ -57,8 +57,7 @@ public class AzureBoardsGlobalUIConfig extends UIConfig {
         super(AzureBoardsDescriptor.AZURE_BOARDS_LABEL, AzureBoardsDescriptor.AZURE_BOARDS_DESCRIPTION, AzureBoardsDescriptor.AZURE_BOARDS_URL);
         this.encryptionValidator = encryptionValidator;
     }
-
-    //FIXME determine which fields will be returned from the OAuth response, and which the user must provide
+    
     @Override
     public List<ConfigField> createFields() {
         //        ConfigField azureBoardsUrlField = new URLInputConfigField(AzureBoardsDescriptor.KEY_AZURE_BOARDS_URL, LABEL_AZURE_BOARDS_URL, DESCRIPTION_AZURE_BOARDS_URL);
@@ -68,8 +67,6 @@ public class AzureBoardsGlobalUIConfig extends UIConfig {
         ConfigField configureOAuth = new OAuthEndpointButtonField(AzureBoardsDescriptor.KEY_OAUTH, LABEL_OAUTH, DESCRIPTION_OAUTH, BUTTON_LABEL_OAUTH)
                                          .applyRequestedDataFieldKey(AzureBoardsDescriptor.KEY_ORGANIZATION_NAME)
                                          .applyRequestedDataFieldKey(AzureBoardsDescriptor.KEY_CLIENT_ID);
-        // TODO requiredRelatedFields aren't validated when pushing endpoint button fields.
-        //.applyRequiredRelatedField(AzureBoardsDescriptor.KEY_CLIENT_ID);
         return List.of(organizationName, clientId, clientSecret, configureOAuth);
     }
